Add a specialized _mint(bytes calldata addresses) function
This is a common scenario, where client code doesn't already have an SSTORE2 pointer but has a packed array of recipients to mint to.

If we rely on the clients to do the straightforward thing, we get something like this:

     function mint(bytes calldata recipients) public returns (uint256) {
        return _mint(SSTORE2.write(recipients));
     }

But this results in:

- 1 copy of recipients from calldata to memory
- 1 copy to prepend it with "00" to prepare the runtime code
- 1 copy to prepend it with the creation code
- then the actual CREATE call
- immediately followed by an SSTORE2 read that will load the entire buffer in memory again

This could lead to a memory expansion of 98kB (1228 * 20 * 4 bytes).

If we optimize for calldata instead and perform the SSTORE2 write ourselves, we can get away with a single memory copy. This leads to a memory expansion of at most 24kB to hold the full SSTORE2 creation code.

Effect on gas:

     test_mint_newcomer_0001() (gas: -1067 (-1.573%))
     test_mint_newcomer_0010() (gas: -1945 (-1.569%))
     test_mint_newcomer_0100() (gas: -10241 (-1.498%))
     test_mint_newcomer_1000() (gas: -103751 (-1.648%))